14 - Abstracted the note file implementation out to a factory for calendar/text notes
